feat(python/main.py): Updated the Langchain implementations
Upgraded the code to get rid of deprecated imports such as VectorQDBA and RetreivalQA.

The QA Chain is now working for the most part. I can't get a response from the server due to a model error. I will try later with a different model to see if it's a model problem. Maybe the current model doesn't support inference endpoints.