feat(python/main.py): Created the RAG Pipeline for the Chatbot
This pipeline uses the <HuggingFaceH4/zephyr-7b-beta> model designed to act as a helpful assistant and it's faster in its responses as compared to <Qwen/Qwen2.5-1.5B-Instruct> that I was using in the HuggingFace Gradio Space.

The RAG implementation allows me to use a simpler and faster model while still maintaining accuracy.